'pageNo'

+++ back-end/models/position.js
@@ -15,14 +15,39 @@ var Position = mongoose.model('positions', new mongoose.Schema({
     formatTime: String,
     companyLogo: String
 }));
-//list接口
-const list = () => {
-    //返回这个集合中的所有字段
-    return Position.find({}).sort({createTime:-1}).
-    then((results) => {
+//listall接口：返回这个集合中的所有数据
+const listall = (query) => {
+    let _query = query || {}// 查询的约定条件
+    return Position.find(_query)
+    .sort({createTime:-1})
+    .then((results) => {
         return results;
-    }).
-    catch((err) => {
+    }).catch((err) => {
+        return false;
+    })
+}
+//list接口：返回某一页（部分）的数据
+const list = async({pageNo = 1,pageSize=10,keyword=''})=>{
+    //搜索关键字
+    let re = new RegExp(keyword,'i');
+    let _query = keyword? {companyName: re}:{}
+    //查询总数据数
+    let all_item = await listall(_query);
+    return  Position.find(query)
+    .sort({createTime:-1})
+    .skip((pageNo-1)*pageSize)  //从哪里开始
+    .limit(~~pageSize)  //截取多少条
+    .then((results) => {
+        return {
+            items:results,
+            pageInfo:{
+                pageNo, //当前是第几页
+                pageSize, //每页多少条数据
+                total:all_item.length, //总数据长度
+                totalPage:Math.ceil(all_item.length/pageSize) //总页数
+            }
+        }
+    }).catch((err) => {
         return false;
     })
 }
@@ -50,7 +75,9 @@ const remove = async({ id })=>{
     then((results)=>{
         results.deleteId=id;
         //点击删除这条数据后，删除保存的图片
-        fs.removeSync(PATH.resolve(__dirname, '../public'+_row.companyLogo))
+        if(_row.companyLogo){
+            fs.removeSync(PATH.resolve(__dirname, '../public'+_row.companyLogo));
+        }
         return results;
     }).
     catch((err) => {
@@ -68,6 +95,13 @@ const listone = ( { id } )=>{
 }
 //update接口数据
 const update = (body)=>{
+    //如果不修改图片那么继续为未修改前的图片
+    if(!body.companyLogo) delete body.companyLogo
+    //当修改图片后，从问价中删除上一张图片
+    if(body.old_img){
+        fs.removeSync(PATH.resolve(__dirname, '../public'+body.old_img));
+    }
+    console.log(body);
     //如果选了重新发布
     if(body.republish){
         let _timestamp = Date.now()
@@ -83,9 +117,10 @@ const update = (body)=>{
     })
 }
 module.exports = {
-    list,
+    listall,
     save,
     remove,
     listone,
-    update
+    update,
+    list
 }
